refactor(mfa): share the standard error callbacks across commands
Five call sites—exec's detached submit, RunCommandWithRetry, websh's
session create, and cp's upload and download—wrote the same four
ErrorHandlerCallbacks fields out by hand, differing only in
RetryOperation. cmd/edit already had them factored into a local
editErrorCallbacks helper, so the shape was settled; it just was not
shared. mfa.ErrorCallbacks(ac, retry) is that helper promoted, and
cmd/edit now calls it too.

The helper cannot live in utils: api/mfa and api/iam both import utils,
so utils importing them back is a cycle. api/mfa is the closest thing to
a home it already had—three of the four callbacks are its own functions,
and it is already the package every one of these commands reaches for.

tunnel, server and workspace still build the struct themselves. tunnel
creates its client only once MFA actually turns out to be needed and
leaves CheckMFACompleted nil, so it takes HandleCommonErrors' legacy
retry path; server needs no username handling; and workspace opens a
workspace-scoped MFA link instead of a server one. Folding any of them
in would change behavior.

No test was added, and none of the paths this helper feeds is covered
today. cmd/edit/edit_test.go injects a fake editDeps and calls runEdit,
so realEditDeps is never constructed; cmd/ftp/cp_test.go covers path
classification and argument parsing only; and nothing calls
RunCommandWithRetry, RunRemoteExec or CreateWebshSession at all. The one
test that drives HandleCommonErrors, utils/error_handler_test.go, builds
the callback struct inline. What pins this change is the compiler, not
the suite: ErrorCallbacks returns a struct literal with no branch of its
own and each site's RetryOperation closure is carried over verbatim.
Covering these paths means giving exec, websh and cp the seam cmd/edit
already has, which is its own change.

Constraint: utils imports no internal package, so the shared helper cannot live there
Rejected: a new pkg/errhandler package — api/mfa already owns three of the four callbacks and every caller already imports it
Rejected: folding tunnel, server and workspace in — their callback sets differ for reasons that are not cosmetic
Confidence: high
Scope-risk: narrow
Reversibility: clean
