+++ CLAUDE.md
@@ -72,7 +72,7 @@ utils/               # Shared utilities (output, prompts, errors, SSH parsing)
 - **Command registration**: All commands are registered in `cmd/root.go` via `RootCmd.AddCommand()`
 - **API layer**: Each `cmd/` package calls corresponding `api/` package for HTTP requests
 - **SSH-like syntax**: `websh`, `exec`, `cp` support `user@host` syntax via `utils.ParseSSHTarget()`
-- **Error handling**: Common errors (MFA required, username required) are handled via `utils.HandleCommonErrors()` with retry callbacks
+- **Error handling**: Common errors (MFA required, username required) are handled via `utils.HandleCommonErrors()` with retry callbacks. `mfa.ErrorCallbacks(ac, retry)` builds the standard callback set for a command that already holds a client—use it instead of writing the four fields out again. Commands whose callbacks genuinely differ construct `utils.ErrorHandlerCallbacks` themselves: `tunnel` creates its client lazily and takes the legacy retry path, and `server`/`workspace` need no username handling or open a workspace-scoped MFA link. `utils` cannot host the helper—`api/mfa` and `api/iam` both import `utils`
 - **Custom flag parsing**: `websh` and `exec` use `DisableFlagParsing: true` and parse flags manually. `exec` supports `--` separator for remote command flags
 - **Shared command execution**: `exec.RunCommandWithRetry()` wraps `event.RunCommand()` + `HandleCommonErrors()` with MFA/retry logic. Used by both `exec` and `websh`
 - **Browser auto-open**: `utils.OpenBrowser()` opens auth URLs with SSH/headless detection, cross-process debounce (`~/.alpacon/.browser_lock`), and `ALPACON_NO_BROWSER` env var opt-out
