+++ cmd/exec/exec.go
@@ -10,7 +10,6 @@ import (
 	"strings"
 
 	"github.com/alpacax/alpacon-cli/api/event"
-	"github.com/alpacax/alpacon-cli/api/iam"
 	"github.com/alpacax/alpacon-cli/api/mfa"
 	"github.com/alpacax/alpacon-cli/client"
 	"github.com/alpacax/alpacon-cli/cmd/worksession"
@@ -155,23 +154,10 @@ func RunRemoteExec(parsed RemoteExecArgs) {
 	if parsed.Detach {
 		resp, err := event.SubmitCommand(alpaconClient, parsed.Server, parsed.Command, parsed.Username, parsed.Groupname, env, workSessionID)
 		if err != nil {
-			err = utils.HandleCommonErrors(err, parsed.Server, utils.ErrorHandlerCallbacks{
-				OnMFARequired: func(srv string) error {
-					return mfa.HandleMFAError(alpaconClient, srv)
-				},
-				OnUsernameRequired: func() error {
-					_, err := iam.HandleUsernameRequired()
-					return err
-				},
-				CheckMFACompleted: func() (bool, error) {
-					return mfa.CheckMFACompletion(alpaconClient)
-				},
-				RefreshToken: alpaconClient.RefreshToken,
-				RetryOperation: func() error {
-					resp, err = event.SubmitCommand(alpaconClient, parsed.Server, parsed.Command, parsed.Username, parsed.Groupname, env, workSessionID)
-					return err
-				},
-			})
+			err = utils.HandleCommonErrors(err, parsed.Server, mfa.ErrorCallbacks(alpaconClient, func() error {
+				resp, err = event.SubmitCommand(alpaconClient, parsed.Server, parsed.Command, parsed.Username, parsed.Groupname, env, workSessionID)
+				return err
+			}))
 		}
 		if err != nil {
 			utils.HandleWorkSessionError(err, "command", parsed.Server, authMethod, workSessionID)
